cleaning db after every test if you include db config in spring boot test classes

+++ src/test/java/pl/bratek20/cookies/test/dbcleaner/DBCleanerTestListener.java
@@ -1,17 +1,17 @@
 package pl.bratek20.cookies.test.dbcleaner;
 
 import org.jetbrains.annotations.NotNull;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.TestContext;
 import org.springframework.test.context.TestExecutionListener;
 
 public class DBCleanerTestListener implements TestExecutionListener {
 
-    @Autowired
-    private DBCleaner dbCleaner;
-
     @Override
     public void afterTestMethod(@NotNull TestContext testContext) {
-        dbCleaner.deleteAllTables();
+        var context = testContext.getApplicationContext();
+        if (context.containsBean("dbCleaner")) {
+            var dbCleaner = context.getBean(DBCleaner.class);
+            dbCleaner.deleteAllTables();
+        }
     }
 }
